+++ src/de/ralleytn/simple/audio/AudioListener.java
@@ -27,7 +27,7 @@ package de.ralleytn.simple.audio;
 /**
  * Listens to audio events.
  * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 1.1.0
+ * @version 2.0.0
  * @since 1.1.0
  */
 @FunctionalInterface