+++ src/de/ralleytn/simple/audio/HeadReader.java
@@ -31,7 +31,7 @@ import java.util.Map;
 /**
  * Is implemented by all classes whith the purpose of reading the head informations from an audio file.
  * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 1.2.0
+ * @version 2.0.0
  * @since 1.2.0
  */
 public interface HeadReader {