+++ src/de/ralleytn/simple/audio/Playable.java
@@ -27,7 +27,7 @@ package de.ralleytn.simple.audio;
 /**
  * Implemented by everything that can be played.
  * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 1.2.0
+ * @version 2.0.0
  * @since 1.1.0
  */
 public interface Playable {