+++ src/de/ralleytn/simple/audio/Recorder.java
@@ -37,7 +37,7 @@ import javax.sound.sampled.TargetDataLine;
 /**
  * Records audio from an input device.
  * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 1.2.2
+ * @version 2.0.0
  * @since 1.0.0
  */
 public class Recorder {