+++ src/module-info.java
@@ -1,0 +1,10 @@
+module de.ralleytn.simple.audio {
+	
+	requires transitive java.desktop;
+	requires de.jarnbjo;
+	requires javazoom;
+	requires javazoom.spi;
+	requires org.tritonus.share;
+	
+	exports de.ralleytn.simple.audio;
+}